@@ -1,16 +1,18 @@
 #!/usr/bin/env node
 
-const fs = require('node:fs');
-const { REST } = require('@discordjs/rest');
-const { Routes } = require('discord-api-types/v9');
-const { clientId, token } = require('./helpers/env');
+import fs from 'node:fs';
+import { REST } from '@discordjs/rest';
+import { Routes } from 'discord-api-types/v9';
+import { clientId, token } from './helpers/env';
 
-const commands = [];
+const commands : any[] = [];
 const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
 
 for (const file of commandFiles) {
-	const command = require(`./commands/${file}`);
-	commands.push(command.data.toJSON());
+	import(`./commands/${file}`)
+		.then(command => {
+			commands.push(command.data.toJSON());
+		});
 }
 
 const rest = new REST({ version: '9' }).setToken(token);
